+++ Cargo.toml
@@ -34,6 +34,7 @@ async-compatibility-layer = { git = "https://github.com/EspressoSystems/async-co
 ] }
 async-std = { version = "1.9.0", features = ["unstable", "attributes"] }
 async-trait = "0.1"
+atomic-option = "0.1"
 bincode = "1.3"
 clap = { version = "4.4", features = ["derive", "env"] }
 commit = { git = "https://github.com/EspressoSystems/commit.git" }
