+++ src/availability.rs
@@ -10,20 +10,38 @@
 // You should have received a copy of the GNU General Public License along with this program. If not,
 // see <https://www.gnu.org/licenses/>.
 
-use crate::{api::load_api, Payload, QueryError, QueryResult};
+//! Queries for HotShot chain state.
+//!
+//! The availability API provides an objective view of the HotShot blockchain. It provides access
+//! only to normative data: that is, data which is agreed upon by all honest consensus nodes and
+//! which is immutable. This means access to core consensus data structures including leaves,
+//! blocks, and headers, where each query is pure and idempotent. This also means that it is
+//! possible for a client to verify all of the information provided by this API, by running a
+//! HotShot light client and downloading the appropriate evidence with each query.
+//!
+//! This API does not provide any queries which represent only the _current_ state of the chain or
+//! may change over time, and it does not provide information for which there is not (yet) agreement
+//! of a supermajority of consensus nodes. For information about the current dynamic state of
+//! consensus and uncommitted state, try the [status](crate::status) API. For information about the
+//! chain which is tabulated by this specific node and not subject to full consensus agreement, try
+//! the [node](crate::node) API.
+
+use crate::{api::load_api, Payload};
 use clap::Args;
 use derive_more::From;
-use futures::{FutureExt, Stream, StreamExt, TryFutureExt, TryStreamExt};
+use futures::{FutureExt, StreamExt, TryFutureExt};
 use hotshot_types::traits::node_implementation::NodeType;
 use serde::{Deserialize, Serialize};
-use snafu::{OptionExt, ResultExt, Snafu};
+use snafu::{OptionExt, Snafu};
 use std::fmt::Display;
 use std::path::PathBuf;
 use tide_disco::{api::ApiError, method::ReadState, Api, RequestError, StatusCode};
 
 pub(crate) mod data_source;
+mod fetch;
 pub(crate) mod query_data;
 pub use data_source::*;
+pub use fetch::Fetch;
 pub use query_data::*;
 
 #[derive(Args, Default)]
@@ -49,32 +67,19 @@ pub enum Error {
     Request {
         source: RequestError,
     },
-    #[snafu(display("error fetching leaf {resource}: {source}"))]
+    #[snafu(display("leaf {resource} missing or not available"))]
     #[from(ignore)]
-    QueryLeaf {
-        source: QueryError,
+    FetchLeaf {
         resource: String,
     },
-    #[snafu(display("error streaming leaves: {source}"))]
+    #[snafu(display("block {resource} missing or not available"))]
     #[from(ignore)]
-    StreamLeaf {
-        source: QueryError,
-    },
-    #[snafu(display("error fetching block {resource}: {source}"))]
-    #[from(ignore)]
-    QueryBlock {
-        source: QueryError,
+    FetchBlock {
         resource: String,
     },
-    #[snafu(display("error streaming blocks: {source}"))]
+    #[snafu(display("transaction {resource} missing or not available"))]
     #[from(ignore)]
-    StreamBlock {
-        source: QueryError,
-    },
-    #[snafu(display("error fetching transaction {resource}: {source}"))]
-    #[from(ignore)]
-    QueryTransaction {
-        source: QueryError,
+    FetchTransaction {
         resource: String,
     },
     #[snafu(display("transaction index {index} out of range for block {height}"))]
@@ -82,18 +87,6 @@ pub enum Error {
     InvalidTransactionIndex {
         height: u64,
         index: u64,
-    },
-    #[snafu(display("unable to open leaf stream at {}: {}", height, reason))]
-    #[from(ignore)]
-    LeafStream {
-        height: usize,
-        reason: String,
-    },
-    #[snafu(display("unable to open block stream at {}: {}", height, reason))]
-    #[from(ignore)]
-    BlockStream {
-        height: usize,
-        reason: String,
     },
     Custom {
         message: String,
@@ -112,13 +105,10 @@ impl Error {
     pub fn status(&self) -> StatusCode {
         match self {
             Self::Request { .. } => StatusCode::BadRequest,
-            Self::QueryLeaf { source, .. }
-            | Self::StreamLeaf { source, .. }
-            | Self::QueryBlock { source, .. }
-            | Self::StreamBlock { source, .. }
-            | Self::QueryTransaction { source, .. } => source.status(),
+            Self::FetchLeaf { .. } | Self::FetchBlock { .. } | Self::FetchTransaction { .. } => {
+                StatusCode::NotFound
+            }
             Self::InvalidTransactionIndex { .. } => StatusCode::NotFound,
-            Self::LeafStream { .. } | Self::BlockStream { .. } => StatusCode::InternalServerError,
             Self::Custom { status, .. } => *status,
         }
     }
@@ -142,7 +132,7 @@ where
                     Some(height) => ResourceId::Number(height),
                     None => ResourceId::Hash(req.blob_param("hash")?),
                 };
-                state.get_leaf(id).await.context(QueryLeafSnafu {
+                state.get_leaf(id).await.context(FetchLeafSnafu {
                     resource: id.to_string(),
                 })
             }
@@ -153,17 +143,7 @@ where
                 let height = req.integer_param("height")?;
                 state
                     .read(|state| {
-                        async move {
-                            handle_stream_errors(
-                                height,
-                                state.subscribe_leaves(height).await,
-                                |height, err| Error::LeafStream {
-                                    height,
-                                    reason: err.to_string(),
-                                },
-                            )
-                        }
-                        .boxed()
+                        async move { Ok(state.subscribe_leaves(height).await.map(Ok)) }.boxed()
                     })
                     .await
             }
@@ -179,7 +159,7 @@ where
                 Ok(state
                     .get_block(id)
                     .await
-                    .context(QueryBlockSnafu {
+                    .context(FetchBlockSnafu {
                         resource: id.to_string(),
                     })?
                     .header()
@@ -190,24 +170,17 @@ where
         .stream("stream_headers", |req, state| {
             async move {
                 let height = req.integer_param("height")?;
-                state
+                Ok(state
                     .read(|state| {
                         async move {
-                            handle_stream_errors(
-                                height,
-                                state
-                                    .subscribe_blocks(height)
-                                    .await
-                                    .map(|blocks| blocks.map_ok(|block| block.header().clone())),
-                                |height, err| Error::BlockStream {
-                                    height,
-                                    reason: err.to_string(),
-                                },
-                            )
+                            state
+                                .subscribe_blocks(height)
+                                .await
+                                .map(|block| Ok(block.header))
                         }
                         .boxed()
                     })
-                    .await
+                    .await)
             }
             .try_flatten_stream()
             .boxed()
@@ -218,7 +191,7 @@ where
                     Some(height) => ResourceId::Number(height),
                     None => ResourceId::Hash(req.blob_param("hash")?),
                 };
-                state.get_block(id).await.context(QueryBlockSnafu {
+                state.get_block(id).await.context(FetchBlockSnafu {
                     resource: id.to_string(),
                 })
             }
@@ -227,21 +200,11 @@ where
         .stream("stream_blocks", |req, state| {
             async move {
                 let height = req.integer_param("height")?;
-                state
+                Ok(state
                     .read(|state| {
-                        async move {
-                            handle_stream_errors(
-                                height,
-                                state.subscribe_blocks(height).await,
-                                |height, err| Error::BlockStream {
-                                    height,
-                                    reason: err.to_string(),
-                                },
-                            )
-                        }
-                        .boxed()
+                        async move { state.subscribe_blocks(height).await.map(Ok) }.boxed()
                     })
-                    .await
+                    .await)
             }
             .try_flatten_stream()
             .boxed()
@@ -250,14 +213,14 @@ where
             async move {
                 let (block, index) = match req.opt_blob_param("hash")? {
                     Some(hash) => state.get_block_with_transaction(hash).await.context(
-                        QueryTransactionSnafu {
+                        FetchTransactionSnafu {
                             resource: hash.to_string(),
                         },
                     )?,
                     None => {
                         let height = req.integer_param("height")?;
                         let id = ResourceId::Number(height);
-                        let block = state.get_block(id).await.context(QueryBlockSnafu {
+                        let block = state.get_block(id).await.context(FetchBlockSnafu {
                             resource: id.to_string(),
                         })?;
                         let i = req.integer_param("index")?;
@@ -280,21 +243,6 @@ where
             .boxed()
         })?;
     Ok(api)
-}
-
-fn handle_stream_errors<T, S, F>(
-    height: usize,
-    stream: QueryResult<S>,
-    map_err: F,
-) -> Result<impl Stream<Item = Result<T, Error>>, Error>
-where
-    S: Stream<Item = QueryResult<T>>,
-    F: Fn(usize, QueryError) -> Error,
-{
-    Ok(stream
-        .map_err(|err| map_err(height, err))?
-        .enumerate()
-        .map(move |(i, res)| res.map_err(|err| map_err(height + i, err))))
 }
 
 #[cfg(test)]
@@ -342,11 +290,7 @@ mod test {
                     }
                 }
                 Err(Error::Availability {
-                    source:
-                        super::Error::QueryBlock {
-                            source: QueryError::NotFound,
-                            ..
-                        },
+                    source: super::Error::FetchBlock { .. },
                 }) => {
                     tracing::info!(
                         "found end of ledger at height {i}, non-empty blocks are {blocks:?}",
