+++ src/data_source/fs.rs
@@ -12,46 +12,19 @@
 
 #![cfg(feature = "file-system-data-source")]
 
-use super::{
-    ledger_log::{Iter, LedgerLog},
-    VersionedDataSource,
-};
-use crate::{
-    availability::{
-        data_source::{
-            AvailabilityDataSource, BlockId, LeafId, ResourceId, UpdateAvailabilityData,
-        },
-        query_data::{
-            BlockHash, BlockQueryData, LeafHash, LeafQueryData, QueryablePayload, TransactionHash,
-            TransactionIndex,
-        },
-    },
-    metrics::PrometheusMetrics,
-    node::{NodeDataSource, UpdateNodeData},
-    status::data_source::StatusDataSource,
-    MissingSnafu, NotFoundSnafu, Payload, QueryResult,
-};
-use async_trait::async_trait;
-use atomic_store::{AtomicStore, AtomicStoreLoader, PersistenceError};
-use commit::Committable;
-use futures::stream::{self, BoxStream, Stream, StreamExt, TryStreamExt};
-use hotshot_types::traits::{node_implementation::NodeType, signature_key::EncodedPublicKey};
-use serde::{de::DeserializeOwned, Serialize};
-use snafu::OptionExt;
-use std::collections::hash_map::{Entry, HashMap};
-use std::hash::Hash;
-use std::ops::{Bound, RangeBounds};
+use super::{storage::FileSystemStorage, FetchingDataSource};
+use crate::{availability::query_data::QueryablePayload, Payload};
+use atomic_store::AtomicStoreLoader;
+use hotshot_types::traits::node_implementation::NodeType;
 use std::path::Path;
-
-const CACHED_LEAVES_COUNT: usize = 100;
-const CACHED_BLOCKS_COUNT: usize = 100;
 
 /// A data source for the APIs provided in this crate, backed by the local file system.
 ///
 /// Synchronization and atomicity of persisted data structures are provided via [`atomic_store`].
-/// The methods [`commit`](Self::commit), [`revert`](Self::revert), and
-/// [`skip_version`](Self::skip_version) of this type can be used to control synchronization in the
-/// underlying [`AtomicStore`].
+/// The methods [`commit`](super::VersionedDataSource::commit),
+/// [`revert`](super::VersionedDataSource::revert), and [`skip_version`](Self::skip_version) of this
+/// type can be used to control synchronization in the underlying
+/// [`AtomicStore`](atomic_store::AtomicStore).
 ///
 /// # Extension and Composition
 ///
@@ -65,11 +38,10 @@ const CACHED_BLOCKS_COUNT: usize = 100;
 /// wrapping it in [`ExtensibleDataSource`](super::ExtensibleDataSource):
 ///
 /// ```
-/// # use atomic_store::PersistenceError;
 /// # use hotshot_query_service::data_source::{ExtensibleDataSource, FileSystemDataSource};
 /// # use hotshot_query_service::testing::mocks::MockTypes as AppTypes;
 /// # use std::path::Path;
-/// # async fn doc(storage_path: &Path) -> Result<(), PersistenceError> {
+/// # async fn doc(storage_path: &Path) -> Result<(), anyhow::Error> {
 /// type AppState = &'static str;
 ///
 /// let data_source: ExtensibleDataSource<FileSystemDataSource<AppTypes>, AppState> =
@@ -90,22 +62,25 @@ const CACHED_BLOCKS_COUNT: usize = 100;
 /// create an aggregate struct containing both [`FileSystemDataSource`] and your additional module
 /// states. A complication arises from how persistent storage is managed: if other modules have
 /// their own persistent state, should the storage of [`FileSystemDataSource`] and the other modules
-/// be completely independent, or synchronized under the control of a single [`AtomicStore`]?
-/// [`FileSystemDataSource`] supports both patterns: when you create it with
-/// [`create`](Self::create) or [`open`](Self::open), it will open its own [`AtomicStore`] and
-/// manage the synchronization of its own storage, independent of any other persistent data it might
-/// be composed with. But when you create it with [`create_with_store`](Self::create_with_store) or
-/// [`open_with_store`](Self::open_with_store), you may ask it to register its persistent data
-/// structures with an existing [`AtomicStoreLoader`]. If you register other modules' persistent
-/// data structures with the same loader, you can create one [`AtomicStore`] that synchronizes all
-/// the persistent data. Note, though, that when you choose to use
+/// be completely independent, or synchronized under the control of a single
+/// [`AtomicStore`](atomic_store::AtomicStore)? [`FileSystemDataSource`] supports both patterns:
+/// when you create it with [`create`](Self::create) or [`open`](Self::open), it will open its own
+/// [`AtomicStore`](atomic_store::AtomicStore) and manage the synchronization of its own storage,
+/// independent of any other persistent data it might be composed with. But when you create it with
 /// [`create_with_store`](Self::create_with_store) or [`open_with_store`](Self::open_with_store),
-/// you become responsible for ensuring that calls to [`AtomicStore::commit_version`] alternate with
-/// calls to [`FileSystemDataSource::commit`] or [`FileSystemDataSource::revert`].
+/// you may ask it to register its persistent data structures with an existing
+/// [`AtomicStoreLoader`]. If you register other modules' persistent data structures with the same
+/// loader, you can create one [`AtomicStore`](atomic_store::AtomicStore) that synchronizes all the
+/// persistent data. Note, though, that when you choose to use
+/// [`create_with_store`](Self::create_with_store) or [`open_with_store`](Self::open_with_store),
+/// you become responsible for ensuring that calls to
+/// [`AtomicStore::commit_version`](atomic_store::AtomicStore::commit_version) alternate with calls
+/// to [`commit`](super::VersionedDataSource::commit) or
+/// [`revert`](super::VersionedDataSource::revert).
 ///
 /// In the following example, we compose HotShot query service modules with other application-
-/// specific modules, using a single top-level [`AtomicStore`] to synchronize all persistent
-/// storage.
+/// specific modules, using a single top-level [`AtomicStore`](atomic_store::AtomicStore) to
+/// synchronize all persistent storage.
 ///
 /// ```
 /// # use async_std::{sync::{Arc, RwLock}, task::spawn};
@@ -164,21 +139,7 @@ const CACHED_BLOCKS_COUNT: usize = 100;
 ///     Ok(app)
 /// }
 /// ```
-#[derive(custom_debug::Debug)]
-pub struct FileSystemDataSource<Types: NodeType>
-where
-    Payload<Types>: QueryablePayload,
-{
-    index_by_leaf_hash: HashMap<LeafHash<Types>, u64>,
-    index_by_block_hash: HashMap<BlockHash<Types>, u64>,
-    index_by_txn_hash: HashMap<TransactionHash<Types>, (u64, TransactionIndex<Types>)>,
-    index_by_proposer_id: HashMap<EncodedPublicKey, Vec<u64>>,
-    #[debug(skip)]
-    top_storage: Option<AtomicStore>,
-    leaf_storage: LedgerLog<LeafQueryData<Types>>,
-    block_storage: LedgerLog<BlockQueryData<Types>>,
-    metrics: PrometheusMetrics,
-}
+pub type FileSystemDataSource<Types> = FetchingDataSource<Types, FileSystemStorage<Types>, ()>;
 
 impl<Types: NodeType> FileSystemDataSource<Types>
 where
@@ -189,11 +150,8 @@ where
     /// If there is already data at `path`, it will be archived.
     ///
     /// The [FileSystemDataSource] will manage its own persistence synchronization.
-    pub async fn create(path: &Path) -> Result<Self, PersistenceError> {
-        let mut loader = AtomicStoreLoader::create(path, "hotshot_data_source")?;
-        let mut data_source = Self::create_with_store(&mut loader).await?;
-        data_source.top_storage = Some(AtomicStore::open(loader)?);
-        Ok(data_source)
+    pub async fn create(path: &Path) -> anyhow::Result<Self> {
+        FetchingDataSource::new(FileSystemStorage::create(path).await?, ()).await
     }
 
     /// Open an existing [FileSystemDataSource] from storage at `path`.
@@ -201,11 +159,8 @@ where
     /// If there is no data at `path`, a new store will be created.
     ///
     /// The [FileSystemDataSource] will manage its own persistence synchronization.
-    pub async fn open(path: &Path) -> Result<Self, PersistenceError> {
-        let mut loader = AtomicStoreLoader::load(path, "hotshot_data_source")?;
-        let mut data_source = Self::open_with_store(&mut loader).await?;
-        data_source.top_storage = Some(AtomicStore::open(loader)?);
-        Ok(data_source)
+    pub async fn open(path: &Path) -> anyhow::Result<Self> {
+        FetchingDataSource::new(FileSystemStorage::open(path).await?, ()).await
     }
 
     /// Create a new [FileSystemDataSource] using a persistent storage loader.
@@ -214,27 +169,10 @@ where
     /// will be archived.
     ///
     /// The [FileSystemDataSource] will register its persistent data structures with `loader`. The
-    /// caller is responsible for creating an [AtomicStore] from `loader` and managing
-    /// synchronization of the store.
-    pub async fn create_with_store(
-        loader: &mut AtomicStoreLoader,
-    ) -> Result<Self, PersistenceError> {
-        let mut ds = Self {
-            index_by_leaf_hash: Default::default(),
-            index_by_block_hash: Default::default(),
-            index_by_txn_hash: Default::default(),
-            index_by_proposer_id: Default::default(),
-            top_storage: None,
-            leaf_storage: LedgerLog::create(loader, "leaves", CACHED_LEAVES_COUNT)?,
-            block_storage: LedgerLog::create(loader, "blocks", CACHED_BLOCKS_COUNT)?,
-            metrics: Default::default(),
-        };
-
-        // HotShot doesn't emit an event for the genesis block, so we need to manually ensure it is
-        // present.
-        ds.insert_genesis().await?;
-
-        Ok(ds)
+    /// caller is responsible for creating an [AtomicStore](atomic_store::AtomicStore) from `loader`
+    /// and managing synchronization of the store.
+    pub async fn create_with_store(loader: &mut AtomicStoreLoader) -> anyhow::Result<Self> {
+        FetchingDataSource::new(FileSystemStorage::create_with_store(loader).await?, ()).await
     }
 
     /// Open an existing [FileSystemDataSource] using a persistent storage loader.
@@ -243,367 +181,25 @@ where
     /// new store will be created.
     ///
     /// The [FileSystemDataSource] will register its persistent data structures with `loader`. The
-    /// caller is responsible for creating an [AtomicStore] from `loader` and managing
-    /// synchronization of the store.
-    pub async fn open_with_store(loader: &mut AtomicStoreLoader) -> Result<Self, PersistenceError> {
-        let leaf_storage =
-            LedgerLog::<LeafQueryData<Types>>::open(loader, "leaves", CACHED_LEAVES_COUNT)?;
-        let block_storage =
-            LedgerLog::<BlockQueryData<Types>>::open(loader, "blocks", CACHED_BLOCKS_COUNT)?;
-
-        let mut index_by_proposer_id = HashMap::new();
-        let mut index_by_block_hash = HashMap::new();
-        let index_by_leaf_hash = leaf_storage
-            .iter()
-            .flatten()
-            .map(|leaf| {
-                index_by_proposer_id
-                    .entry(leaf.proposer())
-                    .or_insert_with(Vec::new)
-                    .push(leaf.height());
-                update_index_by_hash(&mut index_by_block_hash, leaf.block_hash(), leaf.height());
-                (leaf.hash(), leaf.height())
-            })
-            .collect();
-
-        let mut index_by_txn_hash = HashMap::new();
-        for block in block_storage.iter().flatten() {
-            let height = block.height();
-            for (txn_ix, txn) in block.payload().enumerate() {
-                update_index_by_hash(&mut index_by_txn_hash, txn.commit(), (height, txn_ix));
-            }
-        }
-
-        let mut ds = Self {
-            index_by_leaf_hash,
-            index_by_block_hash,
-            index_by_txn_hash,
-            index_by_proposer_id,
-            leaf_storage,
-            block_storage,
-            top_storage: None,
-            metrics: Default::default(),
-        };
-
-        // HotShot doesn't emit an event for the genesis block, so we need to manually ensure it is
-        // present.
-        ds.insert_genesis().await?;
-
-        Ok(ds)
+    /// caller is responsible for creating an [AtomicStore](atomic_store::AtomicStore) from `loader`
+    /// and managing synchronization of the store.
+    pub async fn open_with_store(loader: &mut AtomicStoreLoader) -> anyhow::Result<Self> {
+        FetchingDataSource::new(FileSystemStorage::open_with_store(loader).await?, ()).await
     }
 
     /// Advance the version of the persistent store without committing changes to persistent state.
     ///
-    /// This function is useful when the [AtomicStore] synchronizing storage for this
-    /// [FileSystemDataSource] is being managed by the caller. The caller may want to persist some
-    /// changes to other modules whose state is managed by the same [AtomicStore]. In order to call
-    /// [AtomicStore::commit_version], the version of this [FileSystemDataSource] must be advanced,
-    /// either by [commit](Self::commit) or, if there are no outstanding changes,
+    /// This function is useful when the [AtomicStore](atomic_store::AtomicStore) synchronizing
+    /// storage for this [FileSystemDataSource] is being managed by the caller. The caller may want
+    /// to persist some changes to other modules whose state is managed by the same
+    /// [AtomicStore](atomic_store::AtomicStore). In order to call
+    /// [AtomicStore::commit_version](atomic_store::AtomicStore::commit_version), the version of
+    /// this [FileSystemDataSource] must be advanced, either by
+    /// [commit](super::VersionedDataSource::commit) or, if there are no outstanding changes,
     /// [skip_version](Self::skip_version).
-    pub fn skip_version(&mut self) -> Result<(), PersistenceError> {
-        self.leaf_storage.skip_version()?;
-        self.block_storage.skip_version()?;
-        if let Some(store) = &mut self.top_storage {
-            store.commit_version()?;
-        }
+    pub async fn skip_version(&mut self) -> anyhow::Result<()> {
+        self.storage_mut().await.skip_version()?;
         Ok(())
-    }
-
-    async fn insert_genesis(&mut self) -> Result<(), PersistenceError> {
-        let block_height = StatusDataSource::block_height(self).await.map_err(|err| {
-            PersistenceError::OtherLoad {
-                inner: Box::new(err),
-            }
-        })?;
-        if block_height == 0 {
-            UpdateAvailabilityData::<Types>::insert_leaf(self, LeafQueryData::genesis()).await?;
-            UpdateNodeData::<Types>::insert_leaf(self, LeafQueryData::genesis()).await?;
-            self.insert_block(BlockQueryData::genesis()).await?;
-            self.commit().await?;
-        }
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl<Types: NodeType> VersionedDataSource for FileSystemDataSource<Types>
-where
-    Payload<Types>: QueryablePayload,
-{
-    type Error = PersistenceError;
-
-    /// Commit the current state to persistent storage.
-    ///
-    /// If the [FileSystemDataSource] is managing its own [AtomicStore] (i.e. it was created with
-    /// [create](Self::create) or [open](Self::open)) it will update the global version as well.
-    /// Otherwise, the caller is responsible for calling [AtomicStore::commit_version] after calling
-    /// this function.
-    async fn commit(&mut self) -> Result<(), PersistenceError> {
-        self.leaf_storage.commit_version().await?;
-        self.block_storage.commit_version().await?;
-        if let Some(store) = &mut self.top_storage {
-            store.commit_version()?;
-        }
-        Ok(())
-    }
-
-    /// Revert changes made to persistent storage since the last call to
-    /// [commit](Self::commit).
-    async fn revert(&mut self) {
-        self.leaf_storage.revert_version().unwrap();
-        self.block_storage.revert_version().unwrap();
-    }
-}
-
-async fn range_stream<T>(
-    mut iter: Iter<'_, T>,
-    range: impl RangeBounds<usize>,
-) -> impl '_ + Stream<Item = QueryResult<T>>
-where
-    T: Clone + Serialize + DeserializeOwned + Sync,
-{
-    let start = range.start_bound().cloned();
-    let end = range.end_bound().cloned();
-
-    // Advance the underlying iterator to the start of the range.
-    let pos = match start {
-        Bound::Included(n) => {
-            if n > 0 {
-                iter.nth(n - 1);
-            }
-            n
-        }
-        Bound::Excluded(n) => {
-            iter.nth(n);
-            n + 1
-        }
-        Bound::Unbounded => 0,
-    };
-
-    stream::unfold((iter, end, pos), |(mut iter, end, pos)| async move {
-        // Check if we have reached the end of the range.
-        let reached_end = match end {
-            Bound::Included(n) => pos > n,
-            Bound::Excluded(n) => pos >= n,
-            Bound::Unbounded => false,
-        };
-        if reached_end {
-            return None;
-        }
-        let opt = iter.next()?;
-        Some((opt.context(MissingSnafu), (iter, end, pos + 1)))
-    })
-}
-
-#[async_trait]
-impl<Types: NodeType> AvailabilityDataSource<Types> for FileSystemDataSource<Types>
-where
-    Payload<Types>: QueryablePayload,
-{
-    type LeafStream = BoxStream<'static, QueryResult<LeafQueryData<Types>>>;
-    type BlockStream = BoxStream<'static, QueryResult<BlockQueryData<Types>>>;
-
-    type LeafRange<'a, R> = BoxStream<'a, QueryResult<LeafQueryData<Types>>>
-    where
-        Self: 'a,
-        R: RangeBounds<usize> + Send;
-    type BlockRange<'a, R> = BoxStream<'a, QueryResult<BlockQueryData<Types>>>
-    where
-        Self: 'a,
-        R: RangeBounds<usize> + Send;
-
-    async fn get_leaf<ID>(&self, id: ID) -> QueryResult<LeafQueryData<Types>>
-    where
-        ID: Into<LeafId<Types>> + Send + Sync,
-    {
-        let n = match id.into() {
-            ResourceId::Number(n) => n,
-            ResourceId::Hash(h) => {
-                *self.index_by_leaf_hash.get(&h).context(NotFoundSnafu)? as usize
-            }
-        };
-        self.leaf_storage
-            .iter()
-            .nth(n)
-            .context(NotFoundSnafu)?
-            .context(MissingSnafu)
-    }
-
-    async fn get_block<ID>(&self, id: ID) -> QueryResult<BlockQueryData<Types>>
-    where
-        ID: Into<BlockId<Types>> + Send + Sync,
-    {
-        let n = match id.into() {
-            ResourceId::Number(n) => n,
-            ResourceId::Hash(h) => {
-                *self.index_by_block_hash.get(&h).context(NotFoundSnafu)? as usize
-            }
-        };
-        self.block_storage
-            .iter()
-            .nth(n)
-            .context(NotFoundSnafu)?
-            .context(MissingSnafu)
-    }
-
-    async fn get_leaf_range<R>(&self, range: R) -> QueryResult<Self::LeafRange<'_, R>>
-    where
-        R: RangeBounds<usize> + Send,
-    {
-        Ok(range_stream(self.leaf_storage.iter(), range).await.boxed())
-    }
-
-    async fn get_block_range<R>(&self, range: R) -> QueryResult<Self::BlockRange<'_, R>>
-    where
-        R: RangeBounds<usize> + Send,
-    {
-        Ok(range_stream(self.block_storage.iter(), range).await.boxed())
-    }
-
-    async fn get_block_with_transaction(
-        &self,
-        hash: TransactionHash<Types>,
-    ) -> QueryResult<(BlockQueryData<Types>, TransactionIndex<Types>)> {
-        let (height, ix) = self.index_by_txn_hash.get(&hash).context(NotFoundSnafu)?;
-        let block = self.get_block(*height as usize).await?;
-        Ok((block, ix.clone()))
-    }
-
-    async fn subscribe_leaves(&self, height: usize) -> QueryResult<Self::LeafStream> {
-        Ok(self
-            .leaf_storage
-            .subscribe(height)
-            .context(MissingSnafu)?
-            .map(Ok)
-            .boxed())
-    }
-
-    async fn subscribe_blocks(&self, height: usize) -> QueryResult<Self::BlockStream> {
-        Ok(self
-            .block_storage
-            .subscribe(height)
-            .context(MissingSnafu)?
-            .map(Ok)
-            .boxed())
-    }
-}
-
-#[async_trait]
-impl<Types: NodeType> UpdateAvailabilityData<Types> for FileSystemDataSource<Types>
-where
-    Payload<Types>: QueryablePayload,
-{
-    type Error = PersistenceError;
-
-    async fn insert_leaf(&mut self, leaf: LeafQueryData<Types>) -> Result<(), Self::Error> {
-        self.leaf_storage
-            .insert(leaf.height() as usize, leaf.clone())?;
-        self.index_by_leaf_hash.insert(leaf.hash(), leaf.height());
-        update_index_by_hash(
-            &mut self.index_by_block_hash,
-            leaf.block_hash(),
-            leaf.height(),
-        );
-        Ok(())
-    }
-
-    async fn insert_block(&mut self, block: BlockQueryData<Types>) -> Result<(), Self::Error> {
-        self.block_storage
-            .insert(block.height() as usize, block.clone())?;
-        for (txn_ix, txn) in block.payload().enumerate() {
-            update_index_by_hash(
-                &mut self.index_by_txn_hash,
-                txn.commit(),
-                (block.height(), txn_ix),
-            );
-        }
-        Ok(())
-    }
-}
-
-/// Update an index mapping hashes of objects to their positions in the ledger.
-///
-/// This function will insert the mapping from `hash` to `pos` into `index`, _unless_ there is
-/// already an entry for `hash` at an earlier position in the ledger.
-fn update_index_by_hash<H: Eq + Hash, P: Ord>(index: &mut HashMap<H, P>, hash: H, pos: P) {
-    match index.entry(hash) {
-        Entry::Occupied(mut e) => {
-            if &pos < e.get() {
-                // Overwrite the existing entry if the new object was sequenced first.
-                e.insert(pos);
-            }
-        }
-        Entry::Vacant(e) => {
-            e.insert(pos);
-        }
-    }
-}
-
-#[async_trait]
-impl<Types: NodeType> NodeDataSource<Types> for FileSystemDataSource<Types>
-where
-    Payload<Types>: QueryablePayload,
-{
-    async fn block_height(&self) -> QueryResult<usize> {
-        StatusDataSource::block_height(self).await
-    }
-
-    async fn get_proposals(
-        &self,
-        id: &EncodedPublicKey,
-        limit: Option<usize>,
-    ) -> QueryResult<Vec<LeafQueryData<Types>>> {
-        let all_ids = self
-            .index_by_proposer_id
-            .get(id)
-            .cloned()
-            .unwrap_or_default();
-        let start_from = match limit {
-            Some(count) => all_ids.len().saturating_sub(count),
-            None => 0,
-        };
-        stream::iter(all_ids)
-            .skip(start_from)
-            .then(|height| self.get_leaf(height as usize))
-            .try_collect()
-            .await
-    }
-
-    async fn count_proposals(&self, id: &EncodedPublicKey) -> QueryResult<usize> {
-        Ok(match self.index_by_proposer_id.get(id) {
-            Some(ids) => ids.len(),
-            None => 0,
-        })
-    }
-}
-
-#[async_trait]
-impl<Types: NodeType> UpdateNodeData<Types> for FileSystemDataSource<Types>
-where
-    Payload<Types>: QueryablePayload,
-{
-    type Error = PersistenceError;
-
-    async fn insert_leaf(&mut self, leaf: LeafQueryData<Types>) -> Result<(), Self::Error> {
-        self.index_by_proposer_id
-            .entry(leaf.proposer())
-            .or_default()
-            .push(leaf.height());
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl<Types: NodeType> StatusDataSource for FileSystemDataSource<Types>
-where
-    Payload<Types>: QueryablePayload,
-{
-    async fn block_height(&self) -> QueryResult<usize> {
-        Ok(self.leaf_storage.iter().len())
-    }
-
-    fn metrics(&self) -> &PrometheusMetrics {
-        &self.metrics
     }
 }
 
@@ -611,9 +207,10 @@ where
 mod impl_testable_data_source {
     use super::*;
     use crate::{
-        data_source::UpdateDataSource,
+        data_source::{UpdateDataSource, VersionedDataSource},
         testing::mocks::{DataSourceLifeCycle, MockTypes},
     };
+    use async_trait::async_trait;
     use hotshot::types::Event;
     use tempdir::TempDir;
