+++ src/data_source/notifier.rs
@@ -1,0 +1,362 @@
+// Copyright (c) 2022 Espresso Systems (espressosys.com)
+// This file is part of the HotShot Query Service library.
+//
+// This program is free software: you can redistribute it and/or modify it under the terms of the GNU
+// General Public License as published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+// This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
+// even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+// General Public License for more details.
+// You should have received a copy of the GNU General Public License along with this program. If not,
+// see <https://www.gnu.org/licenses/>.
+
+//! Multi-producer, multi-consumer oneshot notification channel
+//!
+//! [`Notifier`] is an asynchronous, multi-producer, multi-consumer, oneshot channel satisfying
+//! three additional requirements:
+//! 1. Dropped receivers do not consume resources.
+//! 2. Messages are not copied for receivers who don't want them.
+//! 3. Minimal resource contention for concurrent subscriptions.
+//!
+//! ## Dropped receivers do not consume resources
+//!
+//! This requirement is a direct prerequisite of the broader requirement that passive requests for
+//! resources do not consume resources. This is important because in general, passive requests may
+//! be for resources that are not guaranteed to exist, and thus may never terminate. Just like we
+//! avoid spawning a task for passive requests, since it may never complete, we need receivers for
+//! passive requests not to persist beyond the lifetime of the requst, or they may never be closed.
+//!
+//! This requirement is implemented via garbage collection: each time a message is sent, resources
+//! belonging to dropped receivers are cleaned up. Thus, strictly speaking, dropped receivers do
+//! consume resources, but only briefly. There is no need to keep them around until the desired
+//! message is delivered, for example.
+//!
+//! ## Messages are not copied for receivers who don't want them.
+//!
+//! The second requirement simplifies the higher level fetching logic by allowing us to maintain a
+//! single channel for all notifications about a particular resource type, rather than separate
+//! channels for each specific request. Since messages are not copied for all subscribers, but only
+//! for the subscribers interested in a particular message, this simplification becomes nearly
+//! cost-free.
+//!
+//! This requirement is implemented by attaching a predicate to each subscription, which takes a
+//! message by reference. The predicate is checked on the sending side, and the message is only
+//! copied to the subscription if the predicate is satisfied.
+//!
+//! ## Minimal resource contention for concurrent subscriptions.
+//!
+//! This is important because subscriptions are requested in response to read-only client requests,
+//! which are supposed to run in parallel as much as possible. By contrast, notifications are
+//! ususally send from internal server tasks (e.g. the background task that updates the data source
+//! when new blocks are committed). It is less of a problem if these internal tasks contend with
+//! each other, because they are not directly blocking responses to clients, and we have more
+//! control over how and when they acquire shared resources.
+//!
+//! This requirement also empowers us to create a simpler design for the high-level fetching logic.
+//! Specifically, we can reuse the same code for fetching individual resources as we use for
+//! long-lived subscription streams (e.g. `subscribe_blocks` is a thin wrapper around
+//! `get_block_range`). We do not have to worry about adding complex logic to reuse notification
+//! subscriptions for long-lived streams, because subscribing anew for each entry in the stream has
+//! low overhead in terms of contention over shared resources -- the dominant caused in any
+//! concurrent channel, after data copying (see above).
+//!
+//! This further lets us simplify the interface of this channel a bit: since all notifications are
+//! oneshot, consumers deal with futures rather than streams.
+//!
+//! This requirement is satisfied by maintaining the list of subscribers to a [`Notifier`] in a way
+//! that moves most resource contention to message senders, rather than receivers. We make the
+//! assumption that there is less concurrency among senders. In the common case, there is just one
+//! sender: the task monitoring HotShot for new blocks. Occasionally, there may be other tasks
+//! spawned to fetch missing resources and send them through the [`Notifier`], but these should be
+//! relatively few and rare.
+
+use async_compatibility_layer::channel::{
+    oneshot, unbounded, OneShotReceiver, OneShotSender, UnboundedReceiver, UnboundedSender,
+};
+use async_std::sync::{Arc, Weak};
+use atomic_option::AtomicOption;
+use derivative::Derivative;
+use futures::future::{BoxFuture, FutureExt};
+use std::{future::IntoFuture, sync::atomic::Ordering};
+
+pub trait Predicate<T>: 'static + Send + Sync + Fn(&T) -> bool {}
+impl<F, T> Predicate<T> for F where F: 'static + Send + Sync + Fn(&T) -> bool {}
+
+#[derive(Derivative)]
+#[derivative(Debug)]
+struct Subscriber<T> {
+    #[derivative(Debug = "ignore")]
+    predicate: Box<dyn Predicate<T>>,
+    #[derivative(Debug = "ignore")]
+    sender: Arc<AtomicOption<OneShotSender<T>>>,
+}
+
+impl<T> Subscriber<T> {
+    fn is_closed(&self) -> bool {
+        self.sender.load_raw(Ordering::Relaxed).is_null()
+    }
+}
+
+impl<T: Clone> Subscriber<T> {
+    fn notify(&self, msg: &T) {
+        // First, check if the subscriber has been closed. If it has, we can skip it and save the
+        // work of evaluating the predicate.
+        if self.is_closed() {
+            return;
+        }
+        // At this point, it is likely, but not guaranteed, that the subscriber is not closed.
+        // It may have been closed the instant after the check above. However, it is harmless to
+        // evaluate the predicate now; at worst we waste a bit of computation.
+        if !(self.predicate)(msg) {
+            return;
+        }
+        // Now we are committed to sending the message to this subscriber if possible. We can take
+        // the sender. We need to check for closed again in case the subscriber was closed since the
+        // previous check.
+        if let Some(sender) = self.sender.take(Ordering::Relaxed) {
+            sender.send(msg.clone());
+        }
+    }
+}
+
+/// Multi-producer, multi-consumer oneshot notification channel
+#[derive(Derivative)]
+#[derivative(Debug)]
+pub struct Notifier<T> {
+    // Active subscribers.
+    active: Vec<Subscriber<T>>,
+    // Pending subscribers.
+    //
+    // When a new subscriber joins the subscriber set, they do not immediately add themselves to
+    // `active`. Instead, they simply send themselves to this channel. Every time a message is
+    // sent, it will drain pending subscribers from here and add them to `active`. In this way,
+    // almost all of the cost is paid by senders, rather than receivers. We adopt this design for
+    // two reasons:
+    // 1. Most basically, more messages are received than sent, since the intended use of this
+    //    channel is broadcast, with each message being delivered to multiple receivers. Thus,
+    //    moving cost from receivers to senders is always good.
+    // 2. This channel is intended to be highly concurrent, and `active` is a shared resource that
+    //    is ripe for contention. With this design, _only_ sends contend for a lock on the
+    //    subscribers list, but new subscriptions do not. Again, since there are often more
+    //    receivers than senders, this can reduce contention significantly.
+    // Consider the example use case of `Fetcher::fetching_stream`. This design of the notification
+    // channel lets us go with a very simple design of fetching stream, where we fetch each
+    // subsequent entry individually, with a one-shot notification future. This pattern leads to
+    // many concurrent subscription requests: each time a new entry is produced, every open stream
+    // will subscribe anew to notifications for the next entry, at the same time. However, these
+    // concurrent subscriptions go through the high-throughput multi-producer stream implementation,
+    // and do not contend for the lock on `FilterSender::subscribers`.
+    #[derivative(Debug = "ignore")]
+    pending: UnboundedReceiver<Subscriber<T>>,
+    #[derivative(Debug = "ignore")]
+    subscribe: UnboundedSender<Subscriber<T>>,
+}
+
+impl<T> Notifier<T> {
+    pub fn new() -> Self {
+        let (subscribe, pending) = unbounded();
+        Self {
+            active: vec![],
+            pending,
+            subscribe,
+        }
+    }
+}
+
+impl<T: Clone> Notifier<T> {
+    /// Notify all subscribers whose predicate is satisfied by `msg`.
+    pub fn notify(&mut self, msg: &T) {
+        // Try sending the message to each active subscriber.
+        for subscriber in &self.active {
+            subscriber.notify(msg);
+        }
+
+        // Some subscribers may be closed, either because the receiver was dropped or because we
+        // just sent it its message. Remove these from the `active` list.
+        self.active.retain(|subscriber| !subscriber.is_closed());
+
+        // Promote pending subscribers to active and send them the message.
+        for subscriber in self.pending.drain().unwrap_or_default() {
+            subscriber.notify(msg);
+            if !subscriber.is_closed() {
+                // If that message didn't satisfy the subscriber, or it was dropped, at it to the
+                // active list so it will get future messages.
+                self.active.push(subscriber);
+            }
+        }
+    }
+}
+
+impl<T> Notifier<T> {
+    /// Wait for a message satisfying `predicate`.
+    pub async fn wait_for(&self, predicate: impl Predicate<T>) -> WaitFor<T> {
+        // Create a oneshot channel for receiving the notification.
+        let (sender, receiver) = oneshot();
+        let sender = Arc::new(AtomicOption::new(Box::new(sender)));
+
+        // Create a handle which will close the subscription by nulling out the oneshot sender when
+        // dropped. We use a weak referene to handle the case where the notifier is dropped while we
+        // are waiting on the notification. In this case, a strong reference in `handle` would keep
+        // the oneshot sender alive, which would prevent the future from resolving with an error; we
+        // would block forever. The weak reference allows us to close the channel if we are dropped
+        // before the notifier, but also allows the notifier to wake us up by dropping the sender if
+        // it is dropped first.
+        let handle = ReceiveHandle {
+            sender: Arc::downgrade(&sender),
+        };
+
+        // Create a subscriber with our predicate and the oneshot channel.
+        let subscriber = Subscriber {
+            predicate: Box::new(predicate),
+            sender,
+        };
+
+        // Add the subscriber to the channel and return it. We can ignore errors here: `send` only
+        // fails when the receive end of the channel has been dropped, which means the notifier has
+        // been dropped, and thus the send end of the oneshot handle has been dropped. The caller
+        // will discover this when they try to await a notification and get [`None`].
+        self.subscribe.send(subscriber).await.ok();
+        WaitFor { handle, receiver }
+    }
+}
+
+/// A handle that closes a subscriber when dropped.
+struct ReceiveHandle<T> {
+    sender: Weak<AtomicOption<OneShotSender<T>>>,
+}
+
+impl<T> Drop for ReceiveHandle<T> {
+    fn drop(&mut self) {
+        if let Some(sender) = Weak::upgrade(&self.sender) {
+            sender.take(Ordering::Relaxed);
+        }
+    }
+}
+
+/// A pending request for notification.
+///
+/// This object can be `await`ed to block until the requested notification arrives. The result is an
+/// `Option<T>`, which is [`Some`] except in the case that the [`Notifier`] was dropped without ever
+/// sending a satisfying message.
+///
+/// If [`WaitFor`] is dropped before a notification is delivered, it will automatically clean up its
+/// resources in the [`Notifier`].
+pub struct WaitFor<T> {
+    handle: ReceiveHandle<T>,
+    receiver: OneShotReceiver<T>,
+}
+
+impl<T> IntoFuture for WaitFor<T>
+where
+    T: Send + 'static,
+{
+    type Output = Option<T>;
+    type IntoFuture = BoxFuture<'static, Self::Output>;
+
+    fn into_future(self) -> Self::IntoFuture {
+        async move {
+            let res = self.receiver.recv().await.ok();
+
+            // Explicitly drop `handle` _after_ we're done with `receiver`. If the compiler decides
+            // that it can drop `handle` earlier, we might never get a notification.
+            drop(self.handle);
+
+            res
+        }
+        .boxed()
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use crate::testing::setup_test;
+    use async_std::future::timeout;
+    use std::time::Duration;
+
+    #[async_std::test]
+    async fn test_notify_drop() {
+        setup_test();
+        let mut n = Notifier::new();
+
+        // Create two subscribers with different predicates.
+        let wait_for_zero = n.wait_for(|i| *i == 0).await;
+        let wait_for_one = n.wait_for(|i| *i == 1).await;
+
+        // Send a message which satisfies only one of the subscribers.
+        n.notify(&0);
+        assert_eq!(wait_for_zero.await.unwrap(), 0);
+
+        // Check that the other subscriber was not notified.
+        timeout(Duration::from_secs(1), wait_for_one.into_future())
+            .await
+            .unwrap_err();
+
+        // Check subscribers. The first subsciber should have been cleaned up when it was notified.
+        // The second should have been closed when it was dropped without completing, but not yet
+        // garbage collected.
+        assert_eq!(n.active.len(), 1);
+        assert!(n.active[0].is_closed());
+    }
+
+    #[async_std::test]
+    async fn test_notify_active() {
+        setup_test();
+        let mut n = Notifier::new();
+
+        // Create two subscribers.
+        let s1 = n.wait_for(|i| *i == 1).await;
+        let s2 = n.wait_for(|i| *i == 1).await;
+
+        // Send a message that doesn't notify either subscriber, but just promotes them from pending
+        // to active.
+        n.notify(&0);
+        // Check active subscribers.
+        assert_eq!(n.active.len(), 2);
+        assert!(!n.active[0].is_closed());
+        assert!(!n.active[1].is_closed());
+
+        // Drop one of the subscribers, then send another non-satisfying message. This should cause
+        // the dropped subscriber to get garbage collected.
+        drop(s2);
+        n.notify(&0);
+        assert_eq!(n.active.len(), 1);
+        assert!(!n.active[0].is_closed());
+
+        // Satisfy the final subscriber.
+        n.notify(&1);
+        assert_eq!(s1.await.unwrap(), 1);
+    }
+
+    #[async_std::test]
+    async fn test_pending_dropped() {
+        setup_test();
+        let mut n = Notifier::new();
+
+        // Create and immediately drop a pending subscriber.
+        drop(n.wait_for(|_| false).await);
+
+        // Check that the subscriber gets garbage collected on the next notification.
+        n.notify(&0);
+        assert_eq!(n.active.len(), 0);
+    }
+
+    #[async_std::test]
+    async fn test_notifier_dropped() {
+        setup_test();
+
+        let mut n = Notifier::new();
+
+        // Create an active subscriber.
+        let fut1 = n.wait_for(|_| false).await;
+        n.notify(&0);
+
+        // Create a pending subscriber.
+        let fut2 = n.wait_for(|_| false).await;
+
+        // Drop the notifier while both kinds of subscribers are blocked.
+        drop(n);
+        assert_eq!(fut1.await, None);
+        assert_eq!(fut2.await, None);
+    }
+}
