+++ src/data_source/storage/ledger_log.rs
@@ -12,11 +12,9 @@
 
 #![cfg(feature = "file-system-data-source")]
 
-use async_compatibility_layer::async_primitives::broadcast::{channel, BroadcastSender};
 use atomic_store::{
     append_log, load_store::BincodeLoadStore, AppendLog, AtomicStoreLoader, PersistenceError,
 };
-use futures::stream::{self, BoxStream, StreamExt};
 use serde::{de::DeserializeOwned, Serialize};
 use std::collections::VecDeque;
 use std::fmt::Debug;
@@ -34,15 +32,6 @@ pub(crate) struct LedgerLog<T: Serialize + DeserializeOwned> {
     // handled differently and `self.store.iter().len()` does not update until a new version is
     // committed.
     pending_inserts: usize,
-
-    // Send handle for a channel where we stream resource.
-    stream: BroadcastSender<T>,
-    // Because we may receive resource out of order, but `stream` must be ordered, we will not
-    // necessarily send a resource as soon as we get it. `stream_pos` is the index in `store` of the
-    // next object to be sent on `stream` when we receive it. It is equal to the length of the
-    // longest prefix of non-[None] objects in `store`, which may be less than the total length of
-    // `store`.
-    stream_pos: usize,
 }
 
 impl<T: Serialize + DeserializeOwned + Clone> LedgerLog<T> {
@@ -62,8 +51,6 @@ impl<T: Serialize + DeserializeOwned + Clone> LedgerLog<T> {
                 1u64 << 20, // 1 MB
             )?,
             pending_inserts: 0,
-            stream: channel().0,
-            stream_pos: 0,
         })
     }
 
@@ -102,25 +89,12 @@ impl<T: Serialize + DeserializeOwned + Clone> LedgerLog<T> {
             .collect::<VecDeque<_>>();
         cache.reserve_exact(cache_size - cache.len());
 
-        // Find the next object to broadcast on `stream` when it becomes available. This is the
-        // index of the first _unavailable_ item in `store`.
-        let stream_pos = store
-            .iter()
-            .position(|entry| !matches!(entry, Ok(Some(_))))
-            // If there are no currently unavailable entries in `store`, then the next object to
-            // broadcast is the next object to be appended to `store`, whose index is the length of
-            // `store`.
-            .unwrap_or_else(|| store.iter().len());
-        tracing::debug!("stream_pos={}", stream_pos);
-
         Ok(Self {
             cache_start,
             cache_size,
             cache,
             store,
             pending_inserts: 0,
-            stream: channel().0,
-            stream_pos,
         })
     }
 
@@ -150,7 +124,7 @@ impl<T: Serialize + DeserializeOwned + Clone> LedgerLog<T> {
     {
         // If there are missing objects between what we currently have and `object`, pad with
         // placeholders.
-        let len = self.store.iter().len() + self.pending_inserts;
+        let len = self.store.iter().len();
         let target_len = std::cmp::max(index, len);
         for i in len..target_len {
             tracing::debug!("storing placeholders for position {}/{target_len}", len + i);
@@ -170,7 +144,10 @@ impl<T: Serialize + DeserializeOwned + Clone> LedgerLog<T> {
             // This is an object earlier in the chain that we are now receiving asynchronously.
             // Update the placeholder with the actual contents of the object.
             // TODO update persistent storage once AppendLog supports updates.
-            warn!("skipping out-of-order object; random inserts not yet supported");
+            warn!(
+                index,
+                len, target_len, "skipping out-of-order object; random inserts not yet supported"
+            );
 
             // Update the object in cache if necessary.
             if index >= self.cache_start {
@@ -184,18 +161,6 @@ impl<T: Serialize + DeserializeOwned + Clone> LedgerLog<T> {
         tracing::debug!("committing new version of LedgerLog");
         self.store.commit_version()?;
         self.pending_inserts = 0;
-
-        // Broadcast any newly-appended objects which extend the in-order available prefix.
-        let mut i = self.stream_pos;
-        let mut objects = self.iter().skip(i);
-        while let Some(Some(obj)) = objects.next() {
-            tracing::debug!("broadcasting new object {}", i);
-            // Ignore errors on sending, it just means all listeners have dropped their handles.
-            self.stream.send_async(obj).await.ok();
-            i += 1;
-        }
-        self.stream_pos = i;
-
         Ok(())
     }
 
@@ -213,44 +178,6 @@ impl<T: Serialize + DeserializeOwned + Clone> LedgerLog<T> {
 
         self.pending_inserts = 0;
         Ok(())
-    }
-}
-
-impl<T: Serialize + DeserializeOwned + Clone + Send + 'static> LedgerLog<T> {
-    pub(crate) fn subscribe(&self, from: usize) -> Option<BoxStream<'static, T>> {
-        tracing::debug!(
-            "subcribing to objects from {}, stream_pos={}, len={}",
-            from,
-            self.stream_pos,
-            self.iter().len()
-        );
-
-        // A prefix of items are already available, from `from` to `self.stream_pos`. We can yield
-        // these immediately.
-        let prefix = self
-            .iter()
-            .skip(from)
-            // `saturating_sub` handles the case where `from >= self.stream_pos`, in which case
-            // `prefix` is empty.
-            .take(self.stream_pos.saturating_sub(from))
-            .collect::<Option<Vec<_>>>()?;
-        // After the prefix comes the asynchronous stream of items yielded by `self.stream`. Convert
-        // the receive handle for `self.stream` into an instance of `Stream`.
-        let stream = stream::unfold(self.stream.handle_sync(), |mut handle| async move {
-            match handle.recv_async().await {
-                Ok(obj) => Some((obj, handle)),
-                Err(_) => {
-                    // An error in receive means the send end of the channel has been disconnected,
-                    // which means the stream is over.
-                    None
-                }
-            }
-        });
-        // Filter the asynchronous stream so that it starts at `self.stream_pos` or `from`,
-        // whichever comes later.
-        let rest = stream.skip(from.saturating_sub(self.stream_pos));
-        // Return the concatenation of these two streams.
-        Some(stream::iter(prefix).chain(rest).boxed())
     }
 }
 
@@ -411,77 +338,5 @@ mod test {
             );
         }
         assert_eq!(log.iter().nth(5), None);
-    }
-
-    #[async_std::test]
-    async fn test_ledger_log_subscribe() {
-        setup_test();
-
-        let dir = TempDir::new("test_ledger_log").unwrap();
-        let mut loader = AtomicStoreLoader::create(dir.path(), "test_ledger_log").unwrap();
-        let mut log = LedgerLog::<u64>::create(&mut loader, "ledger", 3).unwrap();
-        let mut store = AtomicStore::open(loader).unwrap();
-
-        log.store_resource(Some(0)).unwrap();
-        log.commit_version().await.unwrap();
-        store.commit_version().unwrap();
-
-        // Subscribe one stream starting from items that already exist and one stream starting from
-        // the future.
-        let mut past = log.subscribe(0).unwrap();
-        let mut future = log.subscribe(1).unwrap();
-        assert_eq!(past.next().await.unwrap(), 0);
-
-        // Store a new item, it should be reflected in both streams.
-        log.store_resource(Some(1)).unwrap();
-        log.commit_version().await.unwrap();
-        store.commit_version().unwrap();
-        assert_eq!(past.next().await.unwrap(), 1);
-        assert_eq!(future.next().await.unwrap(), 1);
-
-        // Store two items out of order, they should be reflected in the streams in order.
-        log.insert(3, 3).unwrap();
-        log.commit_version().await.unwrap();
-        store.commit_version().unwrap();
-        log.insert(2, 2).unwrap();
-        log.commit_version().await.unwrap();
-        store.commit_version().unwrap();
-        assert_eq!(past.next().await.unwrap(), 2);
-        assert_eq!(future.next().await.unwrap(), 2);
-        assert_eq!(past.next().await.unwrap(), 3);
-        assert_eq!(future.next().await.unwrap(), 3);
-
-        // TODO enable out-of-order reloading tests once AppendLog supports out-of-order insertion
-        // https://github.com/EspressoSystems/hotshot-query-service/issues/16
-        // Store another item out of order, then reload the log from disk.
-        // log.insert(5, 5).unwrap();
-        // log.commit_version().await.unwrap();
-        // store.commit_version().unwrap();
-        // drop(log);
-        // drop(store);
-        // let mut loader = AtomicStoreLoader::load(dir.path(), "test_ledger_log").unwrap();
-        // let mut log = LedgerLog::<u64>::open(&mut loader, "ledger", 3).unwrap();
-        // let mut store = AtomicStore::open(loader).unwrap();
-
-        // // After reloading from disk, the in-order prefix of objects (0-3) should immediately be
-        // // available.
-        // assert_eq!(
-        //     log.subscribe(0).unwrap().take(4).collect::<Vec<_>>().await,
-        //     vec![0, 1, 2, 3]
-        // );
-
-        // // The remaining items should become available once they can all be yielded in order.
-        // log.insert(4, 4).unwrap();
-        // log.commit_version().await.unwrap();
-        // store.commit_version().unwrap();
-        // assert_eq!(
-        //     log.subscribe(0).unwrap().take(6).collect::<Vec<_>>().await,
-        //     vec![0, 1, 2, 3, 4, 5]
-        // );
-
-        // // Dropping the log should terminate the stream.
-        // let stream = log.subscribe(0).unwrap();
-        // drop(log);
-        // assert_eq!(stream.collect::<Vec<_>>().await, vec![0, 1, 2, 3, 4, 5]);
     }
 }
